Move receipt to own class file
