+++ src/cogs/requests.py
@@ -5,32 +5,8 @@ from datetime import datetime
 import traceback
 
 import os
-import re
 
-class Receipt:
-    category : str
-    requestor : str
-    amount : float
-    date_purchase : datetime
-    description : str
-    submit_time : datetime
-    
-    # TODO: more input validation might be needed
-    def __init__(self, category : str, requestor : str, amount : str, date_purchase : datetime, description : str, submit_time : datetime):
-        self.category = category 
-        self.requestor = requestor
-        self.description = description
-
-        #TODO: Check for non num or decimal point
-        amount = re.sub(r',', "", amount)
-        self.amount = float(amount)
-        
-        #TODO: Check for date format
-        self.date_purchase = datetime.strptime(date_purchase, "%Y-%m-%d").date()  # Convert to date object
-        self.submit_time = submit_time
-
-    def toString(self):
-        return f'{self.submit_time}: {self.requestor} requested ${self.amount} from {self.category} for {self.description}'
+from receipt import Receipt
 
 class Reciept_Modal(discord.ui.Modal, title="Reimbursement Request"):
     category = discord.ui.TextInput(
